feat: Add CreateAuction endpoint to AuctionsController
The code changes add a new endpoint `CreateAuction` to the `AuctionsController` in order to create a new auction. The endpoint accepts a `CreateAuctionDto` object and maps it to an `Auction` entity. The auction is then added to the database and a `Created` response is returned with the created auction details.