feat: Add DeleteAuction endpoint to AuctionsController
The code changes add a new endpoint `DeleteAuction` to the `AuctionsController` in order to delete an auction. The endpoint accepts the `id` of the auction to be deleted and checks if the auction exists. If the auction is found, it is removed from the database and the changes are saved. If the deletion is successful, an `Ok` response is returned.