feat: Add AuctionUpdated and AuctionDeleted using MassTransit/RabbitMQ
This commit implements AuctionUpdated and AuctionDeleted to AuctionService and implements the Outbox pattern for handling messages. The Outbox pattern ensures that MessageBroker messages are saved in the Outbox table and are only sent to the MessageBroker after the DB transaction is committed. This guarantees reliable message delivery and prevents messages from being sent if the DB transaction fails. The necessary code changes are made in the AuctionsController and MappingProfiles files. Also implements the Consumer for these messages in SearchService.