feat: Add client credentials for Postman in IdentityService
This commit adds client credentials for Postman in the IdentityService project. The "postman" client is configured with the necessary scopes and redirect URIs to allow Postman to request a token for development testing purposes. This enables easier testing and integration with the IdentityService using Postman as a client. Documentation for this feature can be found [here](https://docs.duendesoftware.com/identityserver/v7/tokens/password_grant/).