feat: Add Microsoft.AspNetCore.Authentication.JwtBearer package
This commit adds the Microsoft.AspNetCore.Authentication.JwtBearer package to the AuctionService project. The package is included as a reference in the AuctionService.csproj file. This package provides JWT bearer authentication support for the AuctionService, allowing it to authenticate requests using JSON Web Tokens. This is necessary for implementing secure authentication and authorization in the AuctionService.