feat: create reverse proxy GatewayService
This commit updates the appsettings.json and launchSettings.json files in the GatewayService project. In appsettings.json, the ReverseProxy section is added to allow Yarp Nuget package to route requests to microservice apis and allowing for required authentication tokens to pass through.
Also, the LogLevel for Microsoft.AspNetCore is changed from "Warning" to "Information". This will increase the logging level for Microsoft.AspNetCore to provide more detailed information during runtime. In launchSettings.json, the applicationUrl is changed to "http://localhost:6001". This sets the URL where the GatewayService will be hosted during development.