Auction component (#2)
* chore: Update Next.js app layout and add Navbar component

* feat: Create auction Listing and AuctionCard components

Query the gateway /search to load all auctions and display them including images

* feat: Add Countdown Timer to Auction Card

* chore: Add countdown timer to AuctionCard component and update Next.js app layout

* feat: Add CardImage component for AuctionCard images

The CardImage component is added to the AuctionCard component to handle the display of auction images. This improves code organization and allows for better reusability of the image display logic.

* feat: Refactor AuctionCard component to use Auction type

Refactor the AuctionCard component to use the Auction type from the types/index.ts file. This improves type safety and ensures that the correct properties are passed to the component.

* feat: Add flowbite-react dependency and pagination to Auction List

Add the flowbite-react dependency to the package.json file and install it. Also, add pagination functionality to the Auction List component using the Flowbite Pagination component. This improves the user experience by allowing users to navigate through multiple pages of auctions.

* feat: Update pagination in Auction List component

Add flowbite-react dependency and implement pagination functionality in the Auction List component. This allows users to navigate through multiple pages of auctions, improving the overall user experience.