feat: Add login button to EmptyFilter component and customize next-auth SignIn page
This commit adds the ability to log in with a login button in the EmptyFilter component. It also includes the EmptyFilter component, which displays a message when there are no matches for the applied filters in the auction listings. This enhancement improves the user experience by providing a login option