Client authentication (#3)
* feat: Update sorting logic in SearchController and Filters component

This commit updates the sorting logic in the SearchController class to sort by auction end date instead of make when the sort parameter is not provided. It also adds new order buttons for sorting the auctions by alphabetical order, end date, and recently added in the Filters component. These changes improve the functionality and user experience of the search and sorting features in the application.

* feat: Add next-auth dependency for authentication for DuendeIdentityServer6

Add the next-auth dependency to the package.json file and install it. This allows for the implementation of authentication in the web app using Next.js. This enhancement improves the security and user experience of the application.

* feat: create login button functionality

Update the issuer URL in the authOptions object of the route.ts file to use 'http' instead of 'https'. This change ensures that the authentication requests are sent to the correct server URL.

* feat: add ability to view session info during development

* feat: Update authentication callbacks in route.ts

Refactor the authentication callbacks in the route.ts file to include additional logic for handling the token and session data. This change ensures that the user's username and access token are properly stored and used throughout the authentication process. It improves the security and functionality of the authentication feature in the web app.

* feat: Add React Dropdown component for navigation

This commit adds the React Dropdown component to the navigation bar of the web app. The dropdown menu includes options for viewing the user's auctions, auctions won, selling a car, and managing the session (for development purposes). It also includes a sign out option. This enhancement improves the functionality and user experience of the navigation bar.

* feat: Add NextAuth.js middleware for protecting session route

This commit adds NextAuth.js middleware to the web app's middleware.ts file. The middleware is configured to protect the '/session' route, redirecting users to the sign-in page if they are not signed in. This enhancement improves the security of the session route and ensures that only authenticated users can access it.

* feat: Add login button to EmptyFilter component and customize next-auth SignIn page

This commit adds the ability to log in with a login button in the EmptyFilter component. It also includes the EmptyFilter component, which displays a message when there are no matches for the applied filters in the auction listings. This enhancement improves the user experience by providing a login option

* feat: create AuthTest component

This commit creates the AuthTest component for testing authentication functionality. These enhancements improve the user experience and security of the web app.

* feat: Update auctionActions.ts to include token authorization in updateAuctionTest()

This commit updates the auctionActions.ts file to include token authorization in the updateAuctionTest() function. It retrieves the access token using the getTokenWorkaround() function from the authActions.ts file and adds it to the request headers. This enhancement improves the security of the update auction functionality by ensuring that only authenticated users can make updates.