feat: Update SignalR URL in Next.js app
The SignalR URL in the Next.js app was not being overridden by the environment variable in the .env.local file when running in a Docker container. This commit updates the SignalRProvider component to conditionally set the apiUrl based on the NODE_ENV environment variable. In production, the apiUrl is set to 'https://api.carauctionnext.com/notifications', while in other environments, it is set to the value of NEXT_PUBLIC_NOTIFY_URL. This fix ensures that the correct SignalR URL is used in all environments.