Publish to prod local (#4)
* feat: Resolve issue with AuthOptions and env props for prod build

The commit resolves an issue with AuthOptions by modifying the route.ts file in the frontend web app. This change is based on the provided article and fixes the type error related to AuthOptions not being assignable to the expected type. This enhancement improves the functionality of the application by ensuring the correct assignment of AuthOptions in the route.ts file.

* feat: create Nextjs Dockerfile and initial notes

* feat: Update Dockerfile for Next.js app and add web-app to docker-compose

The Dockerfile for the Next.js app has been updated to include the necessary changes for the /app/.next/standalone folder. This ensures that the production image contains all the required files to run the Next.js application.

* feat: update docker-compose to allow client web-app to reach identity server

* add custom network for containers and modify identity-svc with static IP

* feat: Update hosts file for reverse proxy services and add nginx-proxy container to docker-compose

* Adding SSL to reverse proxy gateway

* feat: Update SignalR URL in Next.js app

The SignalR URL in the Next.js app was not being overridden by the environment variable in the .env.local file when running in a Docker container. This commit updates the SignalRProvider component to conditionally set the apiUrl based on the NODE_ENV environment variable. In production, the apiUrl is set to 'https://api.carauctionnext.com/notifications', while in other environments, it is set to the value of NEXT_PUBLIC_NOTIFY_URL. This fix ensures that the correct SignalR URL is used in all environments.

* feat: Install sharp npm package for image processing

* feat: Update ClientApp URL to use HTTPS in docker-compose.yml